+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.7] - 2025-06-03
+
+### Changed
+
+- Renamed `uint32BE` to `toUint32BE`
+- Renamed `uint64BE` to `toUint64BE`
+- Renamed `toB64u` to `toB64U`
+- Renamed `fromB64u` to `fromB64U`
+
 ## [1.0.6] - 2025-06-03
 
 ### Added
@@ -36,6 +45,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `toHex`: Converts a Uint8Array to a hexadecimal string
 - `fromHex`: Converts a hexadecimal string to a Uint8Array with support for odd-length strings and whitespace
+- `uint64BE`: Converts a 64-bit unsigned integer to a Uint8Array in big-endian format
+- `writeUInt32BE`: Writes a 32-bit unsigned integer to a Uint8Array in big-endian format
 
 ## [1.0.1] - 2025-04-11